+++ lib/param_parsing/environment.go
@@ -1,24 +1,44 @@
 //nolint:revive // FIXME: don't use an underscore in package name
 package param_parsing
 
-import "os"
+import (
+	"os"
+	"reflect"
+)
 
 func GetParamsFromEnvironment(params Params) Params {
-	if envArch := os.Getenv("TF_ARCH"); envArch != "" {
-		params.Arch = envArch
-		logger.Debugf("Using architecture from environment variable \"TF_ARCH\": %q", envArch)
-	}
-	if envVersion := os.Getenv("TF_VERSION"); envVersion != "" {
-		params.Version = envVersion
-		logger.Debugf("Using version from environment variable \"TF_VERSION\": %q", envVersion)
-	}
-	if envDefaultVersion := os.Getenv("TF_DEFAULT_VERSION"); envDefaultVersion != "" {
-		params.DefaultVersion = envDefaultVersion
-		logger.Debugf("Using default-version from environment variable \"TF_DEFAULT_VERSION\": %q", envDefaultVersion)
-	}
-	if envProduct := os.Getenv("TF_PRODUCT"); envProduct != "" {
-		params.Product = envProduct
-		logger.Debugf("Using product from environment variable \"TF_PRODUCT\": %q", envProduct)
+	reflectedParams := reflect.ValueOf(&params)
+	for _, envVar := range paramMappings {
+		description := envVar.description
+		env := envVar.env
+		param := envVar.param
+		toml := envVar.toml
+
+		if len(env) == 0 {
+			logger.Errorf("Internal error: environment variable name is empty for parameter %q mapping, skipping assignment", param)
+			continue
+		}
+		if len(param) == 0 {
+			logger.Errorf("Internal error: parameter name is empty for environment variable %q mapping, skipping assignment", env)
+			continue
+		}
+		if len(description) == 0 {
+			description = param
+		}
+
+		paramKey := reflect.Indirect(reflectedParams).FieldByName(param)
+		if paramKey.Kind() != reflect.String {
+			logger.Warnf("Parameter %q is not a string, skipping assignment from environment variable %q", param, env)
+			continue
+		}
+
+		if envVarValue := os.Getenv(env); envVarValue != "" {
+			logger.Debugf("%s (%q) from environment variable %q: %q", description, toml, env, envVarValue)
+			if !paramKey.CanSet() {
+				logger.Warnf("Parameter %q cannot be set, skipping assignment from environment variable %q", param, env)
+			}
+			paramKey.SetString(envVarValue)
+		}
 	}
 	return params
 }