+++ lib/param_parsing/parameters.go
@@ -36,6 +36,24 @@ type Params struct {
 	VersionFlag      bool
 }
 
+type paramMapping struct {
+	description string
+	env         string
+	param       string
+	toml        string
+}
+
+// This is used to automatically instate Environment variables and TOML keys
+var paramMappings = []paramMapping{
+	{param: "Arch", env: "TF_ARCH", toml: "arch", description: "CPU architecture"},
+	{param: "CustomBinaryPath", env: "TF_BINARY_PATH", toml: "bin", description: "Custom binary path"},
+	{param: "DefaultVersion", env: "TF_DEFAULT_VERSION", toml: "default-version", description: "Default version"},
+	{param: "InstallPath", env: "TF_INSTALL_PATH", toml: "install", description: "Custom install path"},
+	{param: "LogLevel", env: "TF_LOG_LEVEL", toml: "log-level", description: "Log level"},
+	{param: "Product", env: "TF_PRODUCT", toml: "product", description: "Product"},
+	{param: "Version", env: "TF_VERSION", toml: "version", description: "Version"},
+}
+
 var logger *slog.Logger
 
 func GetParameters() Params {
@@ -165,22 +183,23 @@ func populateParams(params Params) Params {
 	}
 
 	if isShortRun {
-		logger.Debugf("Resolved CPU architecture: %q", params.Arch)
 		if params.DryRun {
 			logger.Info("[DRY-RUN] No changes will be made")
 		} else {
 			logger.Debugf("Resolved dry-run: %t", params.DryRun)
 		}
+
+		logger.Debugf("Resolved CPU architecture: %q", params.Arch)
 		if params.DefaultVersion != "" {
 			logger.Debugf("Resolved fallback version: %q", params.DefaultVersion)
 		}
-		logger.Debugf("Resolved installation path: %q", filepath.Join(params.InstallPath, lib.InstallDir))
-		logger.Debugf("Resolved installation target: %q", params.CustomBinaryPath)
-		logger.Debugf("Resolved installation version: %q", params.Version)
+		logger.Debugf("Resolved binary path: %q", params.CustomBinaryPath)
+		logger.Debugf("Resolved install path: %q", filepath.Join(params.InstallPath, lib.InstallDir))
+		logger.Debugf("Resolved install version: %q", params.Version)
 		logger.Debugf("Resolved log level: %q", params.LogLevel)
 		logger.Debugf("Resolved mirror URL: %q", params.MirrorURL)
 		logger.Debugf("Resolved product name: %q", params.Product)
-		logger.Debugf("Resolved target directory: %q", params.ChDirPath)
+		logger.Debugf("Resolved working directory: %q", params.ChDirPath)
 	}
 
 	return params