+++ lib/param_parsing/toml_test.go
@@ -2,6 +2,7 @@
 package param_parsing
 
 import (
+	"os"
 	"testing"
 
 	"github.com/warrensbox/terraform-switcher/lib"
@@ -16,6 +17,7 @@ func prepare() Params {
 
 func TestGetParamsTOML_BinaryPath(t *testing.T) {
 	expected := "/usr/local/bin/terraform_from_toml"
+	os.Setenv("BIN_DIR_FROM_TOML", "/usr/local/bin")
 	params := prepare()
 	params, err := getParamsTOML(params)
 	if err != nil {
@@ -24,6 +26,21 @@ func TestGetParamsTOML_BinaryPath(t *testing.T) {
 	if params.CustomBinaryPath != expected {
 		t.Errorf("BinaryPath not matching. Got %v, expected %v", params.CustomBinaryPath, expected)
 	}
+	os.Unsetenv("BIN_DIR_FROM_TOML")
+}
+
+func TestGetParamsTOML_InstallPath(t *testing.T) {
+	expected := "/tmp"
+	os.Setenv("INSTALL_DIR_FROM_TOML", "/tmp")
+	params := prepare()
+	params, err := getParamsTOML(params)
+	if err != nil {
+		t.Fatalf("Got error %v", err)
+	}
+	if params.InstallPath != expected {
+		t.Errorf("InstallPath not matching. Got %q, expected %q", params.InstallPath, expected)
+	}
+	os.Unsetenv("INSTALL_DIR_FROM_TOML")
 }
 
 func TestGetParamsTOML_Version(t *testing.T) {