+++ CHANGELOG.md
@@ -7,6 +7,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org).
 
+## [v1.4.3](https://github.com/warrensbox/terraform-switcher/tree/v1.4.3) - 2025-03-19
+
+[Full Changelog](https://github.com/warrensbox/terraform-switcher/compare/v1.4.2...v1.4.3)
+
+### Added
+
+- feat: Add env vars for `install` and `bin` args [#566](https://github.com/warrensbox/terraform-switcher/pull/566) ([yermulnik](https://github.com/yermulnik))
+- fix: reinstate env var expansion in `bin` and add the same for `install` TOML params [#570](https://github.com/warrensbox/terraform-switcher/pull/570) ([yermulnik](https://github.com/yermulnik))
+
+### Other
+
+- docs: Update CHANGELOG with `v1.4.3` [#569](https://github.com/warrensbox/terraform-switcher/pull/569) ([yermulnik](https://github.com/yermulnik))
+
 ## [v1.4.2](https://github.com/warrensbox/terraform-switcher/tree/v1.4.2) - 2025-03-14
 
 [Full Changelog](https://github.com/warrensbox/terraform-switcher/compare/v1.4.1...v1.4.2)